ensure that saga are not run concurrently, resolve #267
