+++ src/components/FitClassifierDialog/FitClassifierDialog/FitClassifierDialog.tsx
@@ -2,7 +2,7 @@ import { FitClassifierDialogAppBar } from "../FitClassifierDialogAppBar";
 import { OptimizerSettingsListItem } from "../OptimizerSettingsListItem/OptimizerSettingsListItem";
 import { DatasetSettingsListItem } from "../DatasetSettingsListItem/DatasetSettingsListItem";
 import { useDispatch, useSelector } from "react-redux";
-import { classifierSlice, projectSlice } from "../../../store/slices";
+import { classifierSlice } from "../../../store/slices";
 import { Dialog, DialogContent, List } from "@mui/material";
 import { ArchitectureSettingsListItem } from "../ArchitectureSettingsListItem";
 import { PreprocessingSettingsListItem } from "../../PreprocessingSettingsListItem/PreprocessingSettingsListItem";
@@ -10,11 +10,7 @@ import { DialogTransition } from "../../DialogTransition";
 import {
   categorizedImagesSelector,
   compiledSelector,
-  trainingPercentageSelector,
 } from "../../../store/selectors";
-import { ImageType } from "../../../types/ImageType";
-import * as _ from "lodash";
-import { Partition } from "../../../types/Partition";
 import { useEffect, useState } from "react";
 import { TrainingHistoryPlot } from "../TrainingHistoryPlot/TrainingHistoryPlot";
 import { ModelSummaryTable } from "./ModelSummary/ModelSummary";
@@ -22,6 +18,7 @@ import { epochsSelector } from "store/selectors/epochsSelector";
 import { AlertStateType, AlertType, defaultAlert } from "types/AlertStateType";
 import { AlertDialog } from "components/AlertDialog/AlertDialog";
 import { alertStateSelector } from "store/selectors/alertStateSelector";
+import { trainingFlagSelector } from "store/selectors/trainingFlagSelector";
 
 type FitClassifierDialogProps = {
   closeDialog: () => void;
@@ -53,7 +50,7 @@ export const FitClassifierDialog = (props: FitClassifierDialogProps) => {
     { x: number; y: number }[]
   >([]);
 
-  const trainingPercentage = useSelector(trainingPercentageSelector);
+  const currentlyTraining = useSelector(trainingFlagSelector);
   const categorizedImages = useSelector(categorizedImagesSelector);
   const compiledModel = useSelector(compiledSelector);
   const alertState = useSelector(alertStateSelector);
@@ -130,36 +127,9 @@ export const FitClassifierDialog = (props: FitClassifierDialogProps) => {
   };
 
   const onFit = async () => {
-    cleanUpStates();
-
-    //first assign train and val partition to all categorized images
-    const categorizedImagesIds = _.shuffle(categorizedImages).map(
-      (image: ImageType) => {
-        return image.id;
-      }
-    );
-
-    //separate ids into train and val datasets
-    const trainDataLength = Math.round(
-      trainingPercentage * categorizedImagesIds.length
-    );
-    const valDataLength = categorizedImagesIds.length - trainDataLength;
-
-    const trainDataIds = _.take(categorizedImagesIds, trainDataLength);
-    const valDataIds = _.takeRight(categorizedImagesIds, valDataLength);
-
-    dispatch(
-      projectSlice.actions.updateImagesPartition({
-        ids: trainDataIds,
-        partition: Partition.Training,
-      })
-    );
-    dispatch(
-      projectSlice.actions.updateImagesPartition({
-        ids: valDataIds,
-        partition: Partition.Validation,
-      })
-    );
+    if (!currentlyTraining) {
+      cleanUpStates();
+    }
 
     dispatch(
       classifierSlice.actions.fit({
