+++ src/store/sagas/classifier/watchEvaluateSaga.ts
@@ -1,7 +1,7 @@
-import { takeEvery } from "redux-saga/effects";
+import { takeLeading } from "redux-saga/effects";
 import { evaluateSaga } from "./evaluateSaga";
 import { classifierSlice } from "../../slices";
 
 export function* watchEvaluateSaga(): any {
-  yield takeEvery(classifierSlice.actions.evaluate.type, evaluateSaga);
+  yield takeLeading(classifierSlice.actions.evaluate.type, evaluateSaga);
 }
