+++ src/store/sagas/classifier/watchFitSaga.ts
@@ -1,7 +1,7 @@
-import { takeEvery } from "redux-saga/effects";
+import { takeLeading } from "redux-saga/effects";
 import { fitSaga } from "./fitSaga";
 import { classifierSlice } from "../../slices";
 
 export function* watchFitSaga(): any {
-  yield takeEvery(classifierSlice.actions.fit.type, fitSaga);
+  yield takeLeading(classifierSlice.actions.fit.type, fitSaga);
 }
