+++ src/store/sagas/classifier/watchPredictSaga.ts
@@ -1,7 +1,7 @@
-import { takeEvery } from "redux-saga/effects";
+import { takeLeading } from "redux-saga/effects";
 import { classifierSlice } from "../../slices";
 import { predictSaga } from "./predictSaga";
 
 export function* watchPredictSaga(): any {
-  yield takeEvery(classifierSlice.actions.predict.type, predictSaga);
+  yield takeLeading(classifierSlice.actions.predict.type, predictSaga);
 }
